separated out characterdialog and suggestiondialog into their own files.  Also renamed the filenames to use lowercase.
